class_00/pyproject.toml
[project]
name = "class-00"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "nadeem", email = "nadeemsangrasi903@gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "litellm>=1.71.0",
    "openai-agents[litellm]>=0.0.16",
    "python-dotenv>=1.1.0",
]

[project.scripts]
oagent = "class_00.open:start"
lagent = "class_00.lite:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
